+++ Client/src/pages/installed-pack.aspx.js
@@ -61,14 +61,14 @@ render(){
         metaKeywords=""
       >
         <div className="header">
-          <div className="avs-logo">
+          <ImageGQL className="headerBackgroundImage" imageName="installed-pack-header.jpg" style={{position: "absolute", height: "665px"}}/>
+            <div className="header__body">
+            <div className="avs-logo">
           <GatsbyLink className="headerLogo" to="/">
               <LogoWrapper href="/">
               </LogoWrapper>
           </GatsbyLink>
           </div>
-          <ImageGQL className="headerBackgroundImage" imageName="installed-pack-header.jpg" style={{position: "absolute", height: "665px"}}/>
-            <div className="header__body">
               <Text as="h1" className="header__heading installed">{this.props.t("Thank you for installing AVS4YOU Installation Package")}</Text>
                 <div className="landing-block"><div className="landing-one-offer-block"><Text className="text-landing-box-gray">{this.props.t("AVS4YOU Unlimited Subscription")}</Text>
                 <div className="after-text-box"></div>