+++ Client/src/pages/installed-video-remaker.aspx.js
@@ -61,15 +61,16 @@ render(){
         metaKeywords=""
       >
         <div className="header">
-          <div className="avs-logo">
+          <ImageGQL className="headerBackgroundImage" imageName="installed-video-editor-header.jpg" style={{position: "absolute"}}/>
+            <div className="header__body">  
+            <div className="avs-logo">
           <GatsbyLink className="headerLogo" to="/">
               <LogoWrapper href="/">
                   <img src={Logo} alt="avs4you logo"/>
               </LogoWrapper>
           </GatsbyLink>
           </div>
-          <ImageGQL className="headerBackgroundImage" imageName="installed-video-editor-header.jpg" style={{position: "absolute"}}/>
-            <div className="header__body">  <ImageGQL className="installedSuper" imageName="installed-video-editor-super.png" style={{margin: "auto"}}/>
+            <ImageGQL className="installedSuper" imageName="installed-video-editor-super.png" style={{margin: "auto"}}/>
               <Text as="h1" className="header__heading installed">{this.props.t("Thank you for installing AVS Video ReMaker")} <br /> {this.props.t("AVS video remaker")}</Text>
              <table className="header__heading exclusive"><Text as="h2" className="header__heading exclusive"><a href={this.props.t(`${this.state.hrefUnlim}`)} style={{color: "#fff"}}>{this.props.t("Exclusive Offer Only Today")}</a></Text></table>
               <Text as="h3" className="header__subtitle installed" style={{maxWidth:700 + 'px'}}>{this.props.t("Save 70 on the full version and edit your video without limitations")}</Text>